+++ src/app/components/edition/edit-companions/edit-companions.component.ts
@@ -3,33 +3,35 @@ import { AlertInput, AlertOptions, LoadingController } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
 import { Subject, takeUntil } from 'rxjs';
 import { MAGIC_NUMBERS } from 'src/app/constants/number.constants';
+import { CharacterController } from 'src/app/controller/characterController';
+import { cloneCompanions } from 'src/app/controller/follower-companion.controller';
 import { clonePowers, getCultsPowers, getPower, openPowerDetail } from 'src/app/controller/power.controller';
 import { Character } from 'src/app/model/character';
+import { Companion } from 'src/app/model/companion';
 import { Power } from 'src/app/model/power';
 import { CultPowers, PowerData, PowersData } from 'src/app/model/powerData';
-import { getPowersDataConfiguration } from 'src/app/pages/detail/detail.page.configuration.helper';
-import { getPowersData } from 'src/app/pages/detail/detail.page.data.helper';
+import { getCompanionsDataConfiguration, getPowersDataConfiguration } from 'src/app/pages/detail/detail.page.configuration.helper';
+import { getCompanionsData, getPowersData } from 'src/app/pages/detail/detail.page.data.helper';
 import { CharactersService } from 'src/app/services/characters.service';
 import { easyConfirmAlert, openAlert } from 'src/app/utils/alert.utils';
 import { TableDataConfiguration } from '../../table/table.component';
 
 @Component({
-  selector: 'app-edit-powers',
-  templateUrl: './edit-powers.component.html',
-  styleUrls: ['./edit-powers.component.scss'],
+  selector: 'app-edit-companions',
+  templateUrl: './edit-companions.component.html',
+  styleUrls: ['./edit-companions.component.scss'],
 })
-export class EditPowersComponent{
+export class EditCompanionsComponent{
 
   @Output() saveCharacter: EventEmitter<any> = new EventEmitter<any>();
   @Output() exitModal: EventEmitter<any> = new EventEmitter<any>();
 
   public tabledata: any[] =[];
-  public tableConfiguration: TableDataConfiguration = getPowersDataConfiguration(this.translate, this.showPowersDetail);
+  public tableConfiguration: TableDataConfiguration = getCompanionsDataConfiguration(this.translate);
   public character: Character = new Character('Pepe');
-  public powersData: PowersData = new PowersData();
-  public powers: Power[] = [];
-  public initialPowers: Power[] = [];
-  public loading = MAGIC_NUMBERS.N_2;
+  public companions: Companion[] = [];
+  public initialCompanions: Companion[] = [];
+  public loading = MAGIC_NUMBERS.N_1;
 
   private ngUnsubscribe: Subject<void> = new Subject<void>();
 
@@ -38,7 +40,6 @@ export class EditPowersComponent{
     private charactersService: CharactersService,
     private loadingController: LoadingController,
   ) { }
-
 
 
   async ngOnInit() {
@@ -55,12 +56,6 @@ export class EditPowersComponent{
       this.finishLoading(loading);
     });
 
-    this.charactersService.powersData
-      .pipe(takeUntil(this.ngUnsubscribe))
-      .subscribe((data)=> {
-      this.powersData = data;
-      this.finishLoading(loading);
-    });
   }
 
   ngOnDestroy(): void {
@@ -71,12 +66,11 @@ export class EditPowersComponent{
   public removeItem(){
     const alertParams: AlertOptions = {
       header: this.translate.instant('EDIT.REMOVE_POWER'),
-      inputs : this.powers.map((power): AlertInput =>{
-        const powerJ = getPower(this.powersData, power.getName());
+      inputs : this.companions.map((companion): AlertInput =>{
         return {
         type: 'radio',
-        label: powerJ.NAME,
-        value: powerJ.ID
+        label: companion.getName(),
+        value: companion.getName(),
       }}),
       buttons: [
         { text:  this.translate.instant('SHARED.CANCEL'), cssClass: 'alert-secondaryButton' },
@@ -84,18 +78,33 @@ export class EditPowersComponent{
           text:  this.translate.instant('SHARED.DELETE'),
           cssClass: 'alert-primaryButton',
           handler:  (data: any) => {
-            this.powers = this.powers.filter(pow => pow.getName() !== data);
-            this.character.setPowers(this.powers);
-            this.tabledata = getPowersData(this.character, this.powersData, this.translate);
+            this.companions = this.companions.filter(pow => pow.getName() !== data);
+            this.character.setCompanions(this.companions);
+            this.tabledata = getCompanionsData(this.character);
     }}]};
     openAlert(alertParams);
   }
 
   public addItem(){
-    // Add Power 1ST step: Choose cult
-    const inputs: AlertInput[] = getCultsPowers(this.powersData, this.translate).cults.map((cult: CultPowers) => (
-     {type: 'radio', label: cult.cultName, value: cult.cultName }
-    ));
+    const inputs: AlertInput[] = [{
+      type: 'text',
+      name: 'name',
+      label: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.NAME'),
+      placeholder: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.NAME'),
+    },
+    {
+      type: 'text',
+      name: 'player',
+      label: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.PLAYER'),
+      placeholder: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.PLAYER'),
+    },
+    {
+      type: 'text',
+      name: 'cult',
+      label: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.CULT'),
+      placeholder: this.translate.instant('DETAIL_PAGE.BACKGROUND_SEC.CULT'),
+    }
+  ];
    const alertParams: AlertOptions = {
      header: this.translate.instant('EDIT.ADD'),
      inputs : inputs,
@@ -104,32 +113,18 @@ export class EditPowersComponent{
        {
          text:  this.translate.instant('SHARED.OK'),
          cssClass: 'alert-primaryButton',
-         handler:  (cultName: string) => {
-           this.addStep2(cultName);
+         handler:  (data: any) => {
+          const validValues =  CharacterController.isNameValid(data.name) &&
+                              CharacterController.isNameValid(data.player) &&
+                              CharacterController.isNameValid(data.cult);
+          if (!validValues){
+          easyConfirmAlert(this.translate.instant('EDIT.INVALID_FIELDS'), ()=>{return;}, this.translate);
+          }
+          this.character.getCompanions().push(new Companion(data.name, data.player, data.cult));
+          this.companions = this.character.getCompanions();
+          this.tabledata = getCompanionsData(this.character);
        }}]};
        openAlert(alertParams);
- }
-
- public addStep2(cultName: string){
-  // Add Power 2ND step: Choose power
-  const cult = getCultsPowers(this.powersData, this.translate).cults.find(cult => cult.cultName === cultName);
-  const inputs: AlertInput[] = cult?.powers.map((power: PowerData) => (
-    {type: 'radio', label: power.NAME, value: power.ID }
-    )) || [];
-  const alertParams: AlertOptions = {
-    header: this.translate.instant('EDIT.ADD'),
-    inputs : inputs,
-    buttons: [
-      { text:  this.translate.instant('SHARED.CANCEL'), cssClass: 'alert-secondaryButton' },
-      {
-        text:  this.translate.instant('SHARED.OK'),
-        cssClass: 'alert-primaryButton',
-        handler:  (powerId: string) => {
-        this.character.getPowers().push(new Power(powerId));
-        this.powers = this.character.getPowers();
-        this.tabledata = getPowersData(this.character, this.powersData, this.translate);
-      }}]};
-   openAlert(alertParams);
  }
 
   public handleSave(){
@@ -140,7 +135,7 @@ export class EditPowersComponent{
     easyConfirmAlert(
       this.translate.instant('EDIT.MAIN_INFO.CHANGES_LOST'),
       () => {
-        this.character.setPowers(this.initialPowers);
+        this.character.setCompanions(this.initialCompanions);
         this.fetch();
         this.exitModal.emit();
       },
@@ -148,9 +143,9 @@ export class EditPowersComponent{
   }
 
   private fetch(){
-    this.powers = this.character.getPowers();
-    this.tabledata = getPowersData(this.character, this.powersData, this.translate);
-    this.initialPowers= clonePowers(this.powers);
+    this.companions = this.character.getCompanions();
+    this.tabledata = getCompanionsData(this.character);
+    this.initialCompanions= cloneCompanions(this.companions);
 }
 
   private finishLoading(loading: HTMLIonLoadingElement) {
@@ -160,7 +155,4 @@ export class EditPowersComponent{
     }
   }
 
-  private showPowersDetail(row: any){
-    openPowerDetail(row);
-  }
 }