+++ src/app/pages/detail/detail.page.ts
@@ -45,13 +45,23 @@ import { openEquipDetail } from 'src/app/controller/combat.equip.controller';
 import { PowersData } from 'src/app/model/powerData';
 import { openPowerDetail } from 'src/app/controller/power.controller';
 
+enum SECTIONS {
+  GENERAL_SECTION = 'GENERAL_SECTION',
+  SKILLS_SECTION = 'SKILLS_SECTION',
+  COMBAT_SECTION = 'COMBAT_SECTION',
+  GIFTS_SECTION = 'GIFTS_SECTION',
+  BACKGROUND_SECTION = 'BACKGROUND_SECTION',
+}
+
 @Component({
   selector: 'app-detail',
   templateUrl: './detail.page.html',
   styleUrls: ['./detail.page.scss'],
 })
 export class DetailPage implements OnInit, OnDestroy{
-  public section = '1';
+
+  public SECTIONS = SECTIONS;
+  public section = SECTIONS.GENERAL_SECTION;
   public character: Character = new Character('Pepe');
   public characters: Character[] = [];
   public giftData: GiftData = new GiftData();
@@ -107,7 +117,8 @@ export class DetailPage implements OnInit, OnDestroy{
 
   async ngOnInit() {
     const loading: HTMLIonLoadingElement = await this.loadingController.create({
-      message: this.translate.instant('SHARED.LOADING')
+      message: this.translate.instant('SHARED.LOADING'),
+      duration: MAGIC_NUMBERS.N_3000
     });
     await loading.present();
 
@@ -265,7 +276,7 @@ export class DetailPage implements OnInit, OnDestroy{
 
   private switchCaracter(character: Character){
     this.character = character;
-    this.section = '1';
+    this.section = SECTIONS.GENERAL_SECTION;
   }
 
   private saveCharacter(){